// src/interfaces/errors.enum.ts
export enum ErrorCodesEnum {
    userNotFound = 'user-not-found',
    notFound = 'NOT-FOUND',
    manyAttempts = 'MANY-ATTEMPTS',
    wrongPassword = 'WRONG-PASSWORD',
    authWrongPassword = 'wrong-password',
    missingEmail = 'missing-email',
    needLogin = 'Faça o login',
    userMismatch = 'user-mismatch',
    missingVerificationCode = 'missing-verification-code',
    codeExpired = 'code-expired',
    invalidVerificationId = 'invalid-verification-id',
    requiresRecentLogin = 'requires-recent-login',
    notAllowed = 'operation-not-allowed',
    alreadUse = 'email-already-in-use',
    toManyAttempts = 'too-many-requests',
    popUpClosed = 'popup-closed-by-user',
    invalidCode = 'invalid-verification-code',
    existDifferent = 'account-exists-with-different-credential',
}